Convert to ulong

+++ Trs.IntegerMapper/ByteArrayIntegerMapper/ByteArrayMapper.cs
@@ -7,10 +7,10 @@ namespace Trs.IntegerMapper.ByteArrayIntegerMapper
     {
         private readonly ByteEnumerableMapper _byteEnumerableMapper = new ByteEnumerableMapper();
 
-        public uint Map(byte[]? inputValue) => _byteEnumerableMapper.Map(inputValue);
+        public ulong Map(byte[]? inputValue) => _byteEnumerableMapper.Map(inputValue);
 
-        public byte[] ReverseMap(uint mappedValue) => _byteEnumerableMapper.ReverseMap(mappedValue).ToArray();
+        public byte[] ReverseMap(ulong mappedValue) => _byteEnumerableMapper.ReverseMap(mappedValue).ToArray();
 
-        public uint MappedObjectsCount => _byteEnumerableMapper.MappedObjectsCount;
+        public ulong MappedObjectsCount => _byteEnumerableMapper.MappedObjectsCount;
     }
 }
